quera/contest/quera_connect_django/quera_contest/requirements.txt
Django~=3.2.4
django-cors-headers~=3.7.0
django-extensions~=3.1.3
djangorestframework~=3.12.4
